fix: moveWinRate の structured dtype を float32 に揃える
ユーザ判断で streaming 経路との厳密一致を採った．

ColumnarBatch.move_win_rate は float32 なので，structured 側を float16
にすると同じ教師信号が経路によって別の値に丸まる．常駐メモリは float16
案の倍になるが，policy 教師の値が経路で一致することを優先する．

+++ tests/maou/domain/data/test_schema.py
@@ -85,9 +85,10 @@ class TestPreprocessingSchema:
         assert dtype["piecesInHand"].subdtype[0] == np.uint8  # type: ignore[misc,index]
         assert dtype["moveLabel"].subdtype[0] == np.float16  # type: ignore[misc,index]
         assert dtype["moveLabel"].shape == (MOVE_LABELS_NUM,)  # type: ignore[misc]
-        # moveWinRate は moveLabel と同じ形・同じ dtype．
-        # policy 教師 (win_rate / weighted モード) が読む．
-        assert dtype["moveWinRate"].subdtype[0] == np.float16  # type: ignore[misc,index]
+        # moveWinRate は moveLabel と同じ形だが dtype は float32．
+        # streaming 経路 (ColumnarBatch) が float32 を渡すので，
+        # 同じ教師信号が経路によって別の値に丸まらないようにする．
+        assert dtype["moveWinRate"].subdtype[0] == np.float32  # type: ignore[misc,index]
         assert dtype["moveWinRate"].shape == (  # type: ignore[misc]
             MOVE_LABELS_NUM,
         )
